objective solved

+++ memory.py
@@ -23,7 +23,7 @@ class ReplayBuffer:
 		device = torch.device("cuda:0" if torch.cuda.is_available() else "cpu")
 
 		states = torch.from_numpy(np.vstack([e.state for e in experiences if e is not None])).float().to(device)
-		actions = torch.from_numpy(np.vstack([e.action for e in experiences if e is not None])).long().to(device)
+		actions = torch.from_numpy(np.vstack([e.action for e in experiences if e is not None])).float().to(device)
 		rewards = torch.from_numpy(np.vstack([e.reward for e in experiences if e is not None])).float().to(device)
 		next_states = torch.from_numpy(np.vstack([e.next_state for e in experiences if e is not None])).float().to(device)
 		dones = torch.from_numpy(np.vstack([e.done for e in experiences if e is not None]).astype(np.uint8)).float().to(device)
